+++ cognite/client/_api/transformation_schedules.py
@@ -9,13 +9,13 @@ from cognite.client.data_classes import (
     OidcCredentials,
     Transformation,
     TransformationDestination,
+    TransformationFilter,
     TransformationJobBlockade,
     TransformationList,
     TransformationSchedule,
     TransformationScheduleList,
     TransformationScheduleUpdate,
 )
-from cognite.client.data_classes.transformations import TransformationFilter
 
 
 class TransformationSchedulesAPI(APIClient):
@@ -76,7 +76,9 @@ class TransformationSchedulesAPI(APIClient):
         utils._auxiliary.assert_exactly_one_of_id_or_external_id(id, external_id)
         return self._retrieve_multiple(ids=id, external_ids=external_id, wrap_ids=True)
 
-    def list(self, include_public: bool = True, limit: Optional[int] = 25,) -> TransformationList:
+    def list(
+        self, include_public: bool = True, limit: Optional[int] = 25,
+    ) -> TransformationScheduleList:
         """`List all transformation schedules. <https://docs.cognite.com/api/playground/#operation/transformationSchedules>`_
 
         Args:
@@ -85,15 +87,15 @@ class TransformationSchedulesAPI(APIClient):
             limit (int): Limits the number of results to be returned. To retrieve all results use limit=-1, default limit is 25.
 
         Returns:
-            TransformationList: List of transformations
+            TransformationScheduleList: List of schedules
 
         Example:
 
-            List transformations::
+            List schedules::
 
                 >>> from cognite.experimental import CogniteClient
                 >>> c = CogniteClient()
-                >>> transformations_list = c.transformations.list()
+                >>> schedules_list = c.transformations.schedules.list()
         """
         filter = TransformationFilter(include_public=include_public).dump(camel_case=True)