+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changes are grouped as follows
 ### Added
 - `transformations` api client, which allows the creation, deletion, update and retrieval of transformations.
 - `transformations.schedules` api client, which allows the schedule, unschedule and retrieval of recurring runs of a transformation.
+- `transformations.notifications` api client, which allows the creation, deletion and retrieval of transformation email notifications.
 
 ## [2.34.0] - 2021-11-5
 ### Added